referal_system/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
phonenumbers==8.13.50
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
uritemplate==4.1.1
gunicorn==23.0.0
